refactor: rename chart
